Fix 11 audit bugs, publish all benchmark artifacts
Critical:
- Fix PEP 621 license metadata (wheel build was broken)
- Fix 16-bit spike truncation: u32→u64 packing (was dropping events >65K neurons)
- Fix label-based Rust dispatch: use model_name, not user label
- Consume Rust-returned voltages in Python (StateMonitor was stale)

High:
- O(n) spike monitor → O(1) event-native record_event()
- SV emitter TODO stubs now return hard errors instead of placeholder RTL
- README: honest numbers (114 Rust, 72 NetworkRunner, 3051+378 tests, core coverage)

Medium:
- Global np.random.seed → per-network Generator
- requirements.txt synced to v3.13.3
- Chay neuron exp overflow clamped
- Clippy useless_vec fixed in cortical_column.rs

Benchmarks:
- Un-gitignore benchmarks/results/ — all artifacts now published
- Criterion SIMD benchmark (190 Gbit/s popcount, i5-11600K)
- README leads with Rust vs Brian2 speedup (39–202×) instead of micro-benchmark

378 Rust tests pass, 3051 Python tests pass, 0 failures.

+++ engine/src/lib.rs
@@ -333,10 +333,10 @@ impl PyNetworkRunner {
         let dict = PyDict::new(py);
         let spike_counts: Vec<u64> = results.spike_counts.iter().map(|&c| c as u64).collect();
         dict.set_item("spike_counts", spike_counts.into_pyarray(py))?;
-        let spike_data: Vec<Py<PyArray1<u32>>> = results
+        let spike_data: Vec<Py<PyArray1<u64>>> = results
             .spike_data
             .into_iter()
-            .map(|v: Vec<u32>| v.into_pyarray(py).unbind())
+            .map(|v: Vec<u64>| v.into_pyarray(py).unbind())
             .collect();
         dict.set_item("spike_data", spike_data)?;
         let voltages: Vec<Py<PyArray1<f64>>> = results
